--- python/aristotle-glossary/aristotle_glossary/templates/aristotle_glossary/glossary_dialog.html
@@ -1,3 +1,0 @@
-<div>
-    {{ form.as_p }}
-</div>
